refactor: move tests into right package
